+++ test/lib/content/tags/image_test.rb
@@ -1,0 +1,52 @@
+# frozen_string_literal: true
+
+require_relative '../../../test_helper'
+
+class ContentTagsImageTest < ActiveSupport::TestCase
+  delegate :rails_blob_path, to: 'Rails.application.routes.url_helpers'
+
+  setup do
+    @page = occams_cms_pages(:default)
+    @file = occams_cms_files(:default)
+  end
+
+  # -- Tests -------------------------------------------------------------------
+
+  def test_init
+    tag = Occams::Content::Tags::Image.new(context: @page, params: ['my_image'])
+    assert_equal 'my_image', tag.identifier
+    assert_equal 'image', tag.as
+  end
+
+  def test_init_without_identifier
+    message = 'Missing identifier label for image tag'
+    assert_raises Occams::Content::Tag::Error, message do
+      Occams::Content::Tags::Image.new(context: @page)
+    end
+  end
+
+  def test_file
+    tag = Occams::Content::Tags::Image.new(context: @page, params: [@file.label])
+    assert_instance_of Occams::Cms::File, tag.file_record
+
+    tag = Occams::Content::Tags::Image.new(context: @page, params: ['invalid'])
+    assert_nil tag.file_record
+  end
+
+  def test_content
+    tag = Occams::Content::Tags::Image.new(
+      context: @page,
+      params: [@file.label, { 'as' => 'image', 'class' => 'html-class' }]
+    )
+    url = rails_blob_path(tag.file, only_path: true)
+    out = "<img src='#{url}' class='html-class' alt='default file' title='default file'/>"
+    assert_equal out, tag.content
+    assert_equal out, tag.render
+  end
+
+  def test_content_when_not_found
+    tag = Occams::Content::Tags::Image.new(context: @page, params: ['invalid'])
+    assert_equal '', tag.content
+    assert_equal '', tag.render
+  end
+end